+++ src/components/form/date-time/date-time-scroller.component.tsx
@@ -44,7 +44,7 @@ const DateTimeScroller = ({
     dispatcher
   );
 
-  const buttonStyles = 'bsc:bg-transparent bsc:cursor-pointer bsc:focus:outline-hidden bsc:hover:bg-primary-1 bsc:hover:text-white bsc:dark:hover:bg-mono-light-3';
+  const buttonStyles = 'bsc:bg-transparent bsc:cursor-pointer bsc:focus:outline-hidden bsc:hover:bg-primary-1! bsc:hover:text-white! bsc:dark:hover:bg-mono-light-3!';
 
   const defaultTemplate = (_props: DateTimeScrollerTemplateProps, children: TypeOrArray<ReactNode>) => (
     <div className="bc-dt-scroller bsc:flex bsc:w-full bsc:flex-row bsc:items-center bsc:px-2 bsc:py-1">{children}</div>
@@ -59,7 +59,10 @@ const DateTimeScroller = ({
           <BeeSoftIcon icon="chevronLeft" size={IconSize.Regular} />
         </Button>
       </div>
-      <div className="bc-dt-scroller-title bsc:grow bsc:cursor-pointer bsc:text-center bsc:hover:text-primary-2 bsc:dark:hover:text-mono-light-3" onClick={onTitleClicked}>
+      <div
+        className="bc-dt-scroller-title bsc:grow bsc:cursor-pointer bsc:text-center bsc:hover:text-primary-2 bsc:dark:hover:text-mono-light-3"
+        onClick={onTitleClicked}
+      >
         {title}
       </div>
       <div className="bc-dt-scroller-right bsc:shrink bsc:cursor-pointer">